Add files via upload
This project is a implementation of a Car Rental API in Node.js. It uses Express as web framework and MongoDB as database. 

The application allows new users to register and authenticate by using JSON Web Tokens. Registered users are able to rent a car and return it at any time. Admins have additional privileges such as adding new cars to the database or viewing all active rentals.